Extract out SetupFormLegacy.
